services/chat-service/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
numpy==2.2.5
faiss-cpu==1.11.0
redis==5.2.1
httpx==0.28.1
pydantic==2.11.3
